+++ src/opts/container.rs
@@ -201,8 +201,8 @@ where
 
 impl ContainerCreateOpts {
     /// Returns a builder for creating a new container.
-    pub fn builder() -> ContainerOptsBuilder {
-        ContainerOptsBuilder::default()
+    pub fn builder() -> ContainerCreateOptsBuilder {
+        ContainerCreateOptsBuilder::default()
     }
 
     /// Serialize options as a JSON string.
@@ -238,7 +238,7 @@ impl ContainerCreateOpts {
 }
 
 #[derive(Default)]
-pub struct ContainerOptsBuilder {
+pub struct ContainerCreateOptsBuilder {
     name: Option<String>,
     params: HashMap<&'static str, Value>,
 }
@@ -275,8 +275,8 @@ impl FromStr for Protocol {
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
-/// Structure used to expose a port on a container with [`expose`](ContainerOptsBuilder::expose) or
-/// [`publish`](ContainerOptsBuilder::publish).
+/// Structure used to expose a port on a container with [`expose`](ContainerCreateOptsBuilder::expose) or
+/// [`publish`](ContainerCreateOptsBuilder::publish).
 pub struct PublishPort {
     port: u32,
     protocol: Protocol,
@@ -336,7 +336,7 @@ impl ToString for PublishPort {
     }
 }
 
-impl ContainerOptsBuilder {
+impl ContainerCreateOptsBuilder {
     pub fn new(name: impl Into<String>) -> Self {
         Self {
             params: Default::default(),
@@ -660,32 +660,32 @@ mod tests {
     #[test]
     fn create_container_opts() {
         test_case!(
-            ContainerOptsBuilder::new("test_image"),
+            ContainerCreateOptsBuilder::new("test_image"),
             r#"{"HostConfig":{},"Image":"test_image"}"#
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image").env(vec!["foo", "bar"]),
+            ContainerCreateOptsBuilder::new("test_image").env(vec!["foo", "bar"]),
             r#"{"Env":["foo","bar"],"HostConfig":{},"Image":"test_image"}"#
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image").env(&["foo", "bar", "baz"]),
+            ContainerCreateOptsBuilder::new("test_image").env(&["foo", "bar", "baz"]),
             r#"{"Env":["foo","bar","baz"],"HostConfig":{},"Image":"test_image"}"#
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image").env(std::iter::once("test")),
+            ContainerCreateOptsBuilder::new("test_image").env(std::iter::once("test")),
             r#"{"Env":["test"],"HostConfig":{},"Image":"test_image"}"#
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image").user("alice"),
+            ContainerCreateOptsBuilder::new("test_image").user("alice"),
             r#"{"HostConfig":{},"Image":"test_image","User":"alice"}"#
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image")
+            ContainerCreateOptsBuilder::new("test_image")
                 .network_mode("host")
                 .auto_remove(true)
                 .privileged(true),
@@ -693,19 +693,19 @@ mod tests {
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image").expose(PublishPort::tcp(80), 8080),
+            ContainerCreateOptsBuilder::new("test_image").expose(PublishPort::tcp(80), 8080),
             r#"{"ExposedPorts":{"80/tcp":{}},"HostConfig":{"PortBindings":{"80/tcp":[{"HostPort":"8080"}]}},"Image":"test_image"}"#
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image")
+            ContainerCreateOptsBuilder::new("test_image")
                 .expose(PublishPort::udp(80), 8080)
                 .expose(PublishPort::sctp(81), 8081),
             r#"{"ExposedPorts":{"80/udp":{},"81/sctp":{}},"HostConfig":{"PortBindings":{"80/udp":[{"HostPort":"8080"}],"81/sctp":[{"HostPort":"8081"}]}},"Image":"test_image"}"#
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image")
+            ContainerCreateOptsBuilder::new("test_image")
                 .publish(PublishPort::udp(80))
                 .publish(PublishPort::sctp(6969))
                 .publish(PublishPort::tcp(1337)),
@@ -713,22 +713,22 @@ mod tests {
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image").publish_all_ports(),
+            ContainerCreateOptsBuilder::new("test_image").publish_all_ports(),
             r#"{"HostConfig":{"PublishAllPorts":true},"Image":"test_image"}"#
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image").log_driver("fluentd"),
+            ContainerCreateOptsBuilder::new("test_image").log_driver("fluentd"),
             r#"{"HostConfig":{"LogConfig":{"Type":"fluentd"}},"Image":"test_image"}"#
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image").restart_policy("on-failure", 10),
+            ContainerCreateOptsBuilder::new("test_image").restart_policy("on-failure", 10),
             r#"{"HostConfig":{"RestartPolicy":{"MaximumRetryCount":10,"Name":"on-failure"}},"Image":"test_image"}"#
         );
 
         test_case!(
-            ContainerOptsBuilder::new("test_image").restart_policy("always", 0),
+            ContainerCreateOptsBuilder::new("test_image").restart_policy("always", 0),
             r#"{"HostConfig":{"RestartPolicy":{"Name":"always"}},"Image":"test_image"}"#
         );
     }